Ensuring that the content of the posts will always be displayed (#1)
Although the point is to always write each post in Spanish and English, the post may not be available in the chosen language (English maybe 😜 ). Now, when this happens, an alert will be displayed to indicate that although the content is not available in the chosen language, the content will be displayed in the other available language.